+++ src/main/java/com/atoudeft/banque/OperationDepot.java
@@ -1,7 +1,7 @@
 package com.atoudeft.banque;
 
 public class OperationDepot extends Operation{
-    public OperationDepot(TypeOperation typeOperation, int montant) {
+    public OperationDepot(TypeOperation typeOperation, double montant) {
         super(typeOperation, montant);
     }
